+++ src/FantomasVsPackage.cs
@@ -13,7 +13,7 @@ namespace FantomasVs
 
     // DO NOT REMOVE THIS MAGICAL INCANTATION NO MATTER HOW MUCH VS WARNS YOU OF DEPRECATION
     // --------------------------------------------------------------------------------------
-    [InstalledProductRegistration("F# Formatting", "F# source code formatting using Fantomas.", "0.7.3", IconResourceID = 400)]
+    [InstalledProductRegistration("F# Formatting", "F# source code formatting using Fantomas.", "0.9", IconResourceID = 400)]
     // --------------------------------------------------------------------------------------
 
     // Package registration attributes
@@ -34,7 +34,7 @@ namespace FantomasVs
         /// </summary>
         public const string PackageGuidString = "74927147-72e8-4b47-a80d-5568807d6878";
 
-        private static TaskCompletionSource<FantomasVsPackage> _instance = new TaskCompletionSource<FantomasVsPackage>();
+        private static readonly TaskCompletionSource<FantomasVsPackage> _instance = new();
         public static Task<FantomasVsPackage> Instance => _instance.Task;
 
         public FantomasOptionsPage Options => GetDialogPage(typeof(FantomasOptionsPage)) as FantomasOptionsPage ?? new FantomasOptionsPage();