+++ src/FantomasWarmUp.cs
@@ -1,12 +1,11 @@
-﻿using Fantomas;
+﻿extern alias FantomasLatest;
+
+using Fantomas;
+using FantomasLatest::Fantomas;
 using FSharp.Compiler.SourceCodeServices;
 using Microsoft.FSharp.Control;
-using Microsoft.VisualStudio.Text;
 using System;
-using System.Collections.Generic;
 using System.Diagnostics;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace FantomasVs
@@ -60,7 +59,7 @@ shuffle (deck |> List.toArray)
         {
             try
             {
-                var fsasync = CodeFormatter.FormatDocumentAsync("tmp.fs", SourceOrigin.SourceOrigin.NewSourceString(SampleCode), FormatConfig.FormatConfig.Default, FSharpParsingOptions.Default, FSharpChecker.Create(null, null, null, null, null, null, null, null));
+                var fsasync = CodeFormatter.FormatDocumentAsync("tmp.fs", SourceOrigin.SourceOrigin.NewSourceString(SampleCode), FormatConfig.FormatConfig.Default, FSharpParsingOptions.Default, FSharpChecker.Create(null, null, null, null, null, null, null, null, null));
                 var _ = await FSharpAsync.StartAsTask(fsasync, null, null);
             }
             catch (Exception ex)